refactor(vue): split useOrderView into domain composables (#339)
Break the 1753-line useOrderView.js into focused composables so no file
exceeds ~400 lines. The thin orchestrator wires shared refs (order, saving,
editingProduct) and core computeds, delegating to:

- useOrderLoad: order/refs/fields/extra-fields loading + empty-order init
- useOrderSave: save / finalize / create order actions
- useOrderProducts: edit / add / delete product dialog state
- useOrderProductOptions: options table/json editing
- useOrderCostRecalc: cost recalculation + shipping/payment baseline
- useOrderCustomer: customer search + duplicate-customer dialog
- useOrderPluginTabs: built-in + MS3OrderTabsRegistry plugin tab registry

Public orderContext contract and the return shape consumed by OrderView.vue
are preserved unchanged. No behavior/API changes.
